Fixes #26476 - Fetch remote resources from isolated net (#28)
This effectively reverts #26205.

When using env variables for HTTP(S) proxy while fetching
remote resources and there is no proxy available,
the resources will not be downloaded even if host
has a direct connectivity, which seemed like a bug
and led me to introduce #26205 in the first place.

So the usage is: if host has a direct access to remote
resources, do not use the env variables. If host is
in isolated netwok, use the env variables and make
sure your HTTP proxy is correctly configured.